Moving helper functions around, moving Firefox launch/quit out of test cases and into project itself
