+++ iris/tests/experiments/basic_url.py
@@ -3,8 +3,7 @@
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
 
-from api.helpers.general import *
-from api.helpers.keyboard_shortcuts import *
+
 from test_case import *
 
 
@@ -18,12 +17,10 @@ class test(base_test):
 
 
     def run(self):
-        #helper function from "general"
-        firefox_exists()
 
         url = "www.google.com"
         # helper function from "awesome_bar"
-        navigate_to(url)
+        navigate(url)
 
         # image: "Google Search" button
         # details: below the search field in page content
@@ -31,7 +28,7 @@ class test(base_test):
         image = "google_search.png"
 
         # core api function
-        if exists(image, 1):
+        if exists(image, 10):
             result = "PASS"
         else:
             result = "FAIL"
